mobile-elements-detector/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mobile-elements-detector"
version = "0.1.0"
description = "A bioinformatics tool for detecting mobile elements and analyzing genomic architecture."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Joao Vito", email = "your.email@example.com"}
]
dependencies = [
    "biopython>=1.81",
    "matplotlib>=3.7.1",
    "rich>=13.3.5",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mobile-elements-detector = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
